fixing https://support.wso2.com/jira/browse/BNYMAPIMDEV-204

+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/APIProviderImpl.java
@@ -1250,7 +1250,9 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
                         }
                     } else { // API Status : RETIRED
                         Map<String, String> failedToRemoveEnvironments = removeFromGateway(api);
-                        apiMgtDAO.removeAllSubscriptions(api.getId());
+                        if (!APIStatus.CREATED.equals(status)) {
+                            apiMgtDAO.removeAllSubscriptions(api.getId());
+                        }
                         if (!failedToRemoveEnvironments.isEmpty()) {
                             Set<String> publishedEnvironments = new HashSet<String>(api.getEnvironments());
                             publishedEnvironments.addAll(failedToRemoveEnvironments.keySet());
@@ -1395,7 +1397,9 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
                             }
                         } else { // API Status : RETIRED or CREATED
                             Map<String, String> failedToRemoveEnvironments = failedGatewaysMap;
-                            apiMgtDAO.removeAllSubscriptions(api.getId());
+                            if(!APIStatus.CREATED.equals(newStatus)) {
+                                apiMgtDAO.removeAllSubscriptions(api.getId());
+                            }
                             if (!failedToRemoveEnvironments.isEmpty()) {
                                 Set<String> publishedEnvironments = new HashSet<String>(api.getEnvironments());
                                 publishedEnvironments.addAll(failedToRemoveEnvironments.keySet());
